fix next button being greyed out when accept aup button was cliked.
As of jQuery 1.6, the .attr() method returns undefined for
attributes that have not been set. This doesn't match the
tests for false in the JavaScript if statements. When this is
changed, the code works again as expected.

The use of 'true' and 'false' in setting the attributes is something
that does keep working for boolean attributes in jQuery 1.6.1.

Signed-off-by: Henrik Austad <henrik@austad.us>