about_nren/privacy: extend test
Testing for nren_unset_(pt|about)_text alone is not sufficient as the
translation framework for about/privacy/help is set to use index by
default and thus, all these messages will be exported to smarty.

The email is one variable which is only set when the user is AuthN, but
the text is not set.

Signed-off-by: Henrik Austad <henrik@austad.us>