Remove sanitizeIdPName from system
We should allow more characters in the name of the IdP. As this is not a
value we use in the certificate itself, we need only clean out
html-values to stop XSS from happening.

Since we use htmlentities() for this, we can remove all references to
sanitizeIdPName() including the function itself. It is important to
remove this from admin.php as it could be difficult to add admins to
Subscribers with invalid names.

Reported-By: Thijs Kinkhorst <thijs@uvt.nl>
Signed-off-by: Henrik Austad <henrik@austad.us>