admin: do not combine truncate and escape for names with special chars
sub-admin pages must use the proper function for retrieving the name,
and in the legend, truncate should be used to avoid lengthy names to
mess with the layout.

When we have special characters, truncate and escape (functions of
smarty) do not play together nicely. The orgname is one value that we
sanitize well, so we can drop escape in the legend of these fields.

Signed-off-by: Henrik Austad <henrik@austad.us>